Fix abigen with int expression template arguments
In my contracts, I use several template arguments which are expressions
which resolve to integral values (such as an eosio::name) but are
not literal integer values. This works great in my contracts, but
causes abigen to fail.

Fix abigen in cases where a template argument is an expression that
resolves to an integer, but isn't an integer literal.