[Refl] Add TypeList, eliminate use of std::tuple
Using std::tuple in reflection logic is inelegant, as tuple is a
value-semantics type rather than a pure compile-time construct, and
these value semantics are vestigial and problematic in compile-time
logic. Tuple also does not provide us with useful compile-time
manipulations of the types being handled.

Use TypeList instead, which is ideal for this kind of processing.